[FLINK-12070] [network] (part 2) Make "release parent releases readers" test specific to pipelined partitions.
For pipelined partitions, the release call on the SubPartition causes immedately releasing the reader (view).
For bounded partitions, this is not required or even desirable, because too eager release can segfault in case
of direct byte buffers and memory mapped files.
